DUOS-1127 split up NewChairConsole components, currently testing functionality
