+++ .eslintrc
@@ -15,7 +15,7 @@
     "ecmaVersion": 2020,
     "sourceType": "module"
   },
-  "plugins": ["@babel"],
+  "plugins": ["@babel", "react-hooks"],
   "rules": {
     "no-trailing-spaces": "warn",
     "no-unreachable": "warn",
@@ -23,6 +23,8 @@
     "array-bracket-spacing": "warn",
     "no-unused-vars": "warn",
     "no-extra-semi": "warn",
+    "react-hooks/rules-of-hooks": "error", // Checks rules of Hooks
+    "react-hooks/exhaustive-deps": "warn", // Checks effect dependencies
     "react/jsx-filename-extension": [
       1,
       {
